refactor(impl): reuse shared helpers, drop dead code, cut hot-path waste
A review pass over `impl/` for reuse, simplification, efficiency and
altitude. Behavior-preserving throughout; the full suite is unchanged at
1951 passed / 2 skipped, and mypy, ty and all 11 import contracts stay green.

Reuse:
- render/target.py: three hand-inlined clip-box intersections now call
  `internal_intersect_box`, the helper ten other sites already use.
- s_08_graphics/color.py: four inlined float->byte quantizations now call
  `internal_component_byte`, already imported from device_profiles.
- parse/layout.py: five identical nine-line recursive forwarding blocks in
  `internal_xy_cut_regions` collapse into one local `recurse()`.
- layout/text_lines.py: four copies of the short-digit-run guard become
  `internal_is_short_digit_run(max_length=...)`.

Simplification:
- layout/lines.py: `LayoutLine.__init__` took 13 geometry parameters and used
  a 14-conjunct sentinel to detect that none were passed. No caller ever
  passed them; it now takes `runs`.
- operator_tables.py: drop the `category` field, its `OperatorCategory` alias
  and 45 `category=` arguments -- nothing ever read it.
- Delete unreachable code: `page_to_json_dict`/`element_to_json_dict`, the two
  `clone()` methods, `RectBox`'s unused operator API, and four dead branches
  (one OCR skip was strictly subsumed by the one above it; one fusion branch
  was byte-identical to its fallthrough).

Efficiency:
- structured/model.py: `DocumentTextView.words` rebuilt every word to stamp a
  page number `TextView.words` had already written -- a provable no-op.
- render/kernels.py: `astype(uint32)` materialized a full copy before
  `reduceat`; passing `dtype=` is bit-identical.
- render/target.py: `paint_shading` and the mask blit ran a `bisect_left` per
  pixel to answer a per-row question, and re-resolved the soft-mask alpha per
  pixel. Both now walk the row spans directly.
- render/target.py: the generation-keyed clip cache was never cleared though
  every entry becomes unreachable when the generation bumps; the path-keyed
  one was unbounded. Both now follow the file's existing cap convention.
- s_07_security/standard.py: memoize the per-object key, which was re-deriving
  an MD5 once per decrypted string rather than once per object.
- font_program_truetype.py: cache normalized contours like the OpenType
  sibling; real programs have 2048 upem, so the rescale never hit its
  `scale == 1.0` shortcut and rebuilt every contour per painted glyph.
- s_09_fonts/decoder.py: hoist the per-glyph call and attribute loads out of
  `text_advance_vector`, preserving operation order so results stay
  bit-identical, and move `bytes(data)` under the branch that uses it.
- s_08_graphics/color.py: the Separation LUT cache keyed on the `PdfStream`
  object, pinning its memoryview into the document buffer for the life of the
  process. It is now keyed by value, so the buffer dies with the document.

Altitude:
- s_07_content/operations.py: the pdfminer-compat flag was resolved by a
  double getattr chain on every `<<` token of every content stream, for a
  value fixed when the document is opened. Hoisted out of the loop.

tests/test_rendering.py asserted on `lru_cache` internals for the Separation
LUT; it now checks the replacement cache and additionally pins the leak fix.

+++ src/core_pdf/impl/render/kernels.py
@@ -592,8 +592,8 @@ def internal_box_downsample(
     column_edges = (numpy.arange(target_width + 1, dtype=numpy.int64) * source_width) // (
         target_width
     )
-    totals = numpy.add.reduceat(grid.astype(numpy.uint32), row_edges[:-1], axis=0)
-    totals = numpy.add.reduceat(totals, column_edges[:-1], axis=1)
+    totals = numpy.add.reduceat(grid, row_edges[:-1], axis=0, dtype=numpy.uint32)
+    totals = numpy.add.reduceat(totals, column_edges[:-1], axis=1, dtype=numpy.uint32)
     counts = numpy.diff(row_edges)[:, None, None] * numpy.diff(column_edges)[None, :, None]
     reduced = (totals // numpy.maximum(counts, 1)).astype(numpy.uint8)
     return reduced.reshape(-1), target_width, target_height
